divide files, fix invalid example range
